requirements.txt
corus

datasets
transformers
seqeval

streamlit
st-annotated-text